refactor(config): extract the field validators
guard_core/_security_config_validators.py was rank B (MI 18.73). The
per-field revalidators (whitelist/blacklist/trusted_proxies IP parsing,
country-list and geo-IP-handler resolution, threat-ban and behavior-rule
config validation, the /0-network and empty-categories warnings) move to
_security_config_field_validators.py.

_revalidate_copied_config split into five named helpers, one per
independent branch: _revalidate_exclude_paths_field,
_revalidate_country_shadow_fields, _revalidate_global_behavior_rule_fields,
_revalidate_geo_state_fields, _revalidate_changed_fields. Its own
cyclomatic complexity drops from 8 to 1 (a straight-line sequence of five
calls); the two warnings it triggers gained a stack frame, so their
stacklevel moves from 4 to 5, verified live against the exact call-site
line through SecurityConfig.model_copy.

cloud_blocking_enabled and _extra_installed move to a new
guard_core/_config_capabilities.py, next to models.py, their only caller;
_security_config_validators.py re-exports both for parity. Neither was
kept in _security_config_validators.py as xenon filler this time, the
module's own average cyclomatic complexity (2.17) clears xenon's
--max-modules A without needing dilution.

_security_config_validators.py re-exports every name it exported before.
scripts/unasync.py's SKIP_SRC gets both new sibling modules (shared,
synchronous validation code with no sync mirror, same as
_security_config_validators.py and _config_field_revalidators.py).

CHANGELOG: one line under Changed for the whole Task 4 module split
(suspatterns_handler.py, handler_initializer.py,
_security_config_validators.py, _redos_structure.py).
