+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Unreleased
 - **`SemanticAnalyzer.analyze` no longer tokenizes its input twice.** `analyze_attack_probability` and the `token_count` field both called `extract_tokens` independently on the same content; `analyze` now extracts tokens once and reuses them for both. Second of the three constant-factor cuts.
 - **The url-decoded detection view no longer re-runs percent/unicode decoding from scratch.** `_check_url_decoded_view_patterns` called `preprocess_url_decoded_newline_preserving`, which independently repeated `decode_common_encodings` on the raw content even when the main preprocessing pass had already decoded it, or found nothing to decode; it now reuses the main pass's decoded intermediate (`ContentPreprocessor.preprocess_with_decoded`) instead of decoding twice. Third of the three constant-factor cuts.
 - **Breaking: the rate limiter now honors `redis_fail_open` on a Redis failure instead of always falling back to the in-memory window.** `_redis_request_count` (`check_rate_limit` and `check_rate_limit_by_ip`, `guard_core/handlers/ratelimit_handler.py`) previously caught every Redis error itself and silently used the in-memory window regardless of `redis_fail_open`, so a Redis outage silently turned a shared, cross-worker rate limit into a per-process one (N workers effectively multiplying `rate_limit` by N) with no way to opt out. `redis_fail_open=True` still falls back to the in-memory window, but now logs a `WARNING` once per process ("Redis unavailable for rate limiting...") instead of an `ERROR` on every request. `redis_fail_open=False` (the default) now raises `GuardRedisError` instead of silently falling back: `check_rate_limit` lets `SecurityCheckPipeline._handle_check_error` apply `fail_secure` exactly as for any other check's Redis failure (500 under `fail_secure=True`, pass-through otherwise), and `check_rate_limit_by_ip` lets the exception reach its own caller. A deployment relying on the previous always-fall-back behavior with the default `redis_fail_open=False` now needs to set `redis_fail_open=True` explicitly to keep it. `NoScriptError` reload handling and the auto-ban feed are unchanged.
+- **Internal module split, no public name changes.** `guard_core/handlers/suspatterns_handler.py`, `guard_core/core/initialization/handler_initializer.py`, `guard_core/_security_config_validators.py` and `guard_core/detection_engine/_redos_structure.py` were each at maintainability rank below A; every one is restructured into cohesive private sibling modules (`guard_core/handlers/_suspatterns_*.py`, `guard_core/detection_engine/_redos_structure_*.py`, and similarly for the other two), with every previously importable name re-exported from its original module path. Detection results, sync mirrors, and behavior are unchanged.
 
 ### Fixed
