+++ scripts/unasync.py
@@ -17,6 +17,8 @@ SKIP_SRC = {
     "exceptions.py",
     "_pydantic_plugin_mute.py",
     "_security_config_validators.py",
+    "_security_config_field_validators.py",
+    "_config_capabilities.py",
     "_config_field_revalidators.py",
     "_security_config_fields.py",
     "_dynamic_rules.py",
